requirements.txt
# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# AI & Text Generation
anthropic==0.18.1
openai==1.12.0

# Text to Speech
elevenlabs==1.1.2

# Video Generation
moviepy==1.0.3
Pillow==10.2.0

# Audio Processing
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
rich==13.7.0
pyyaml==6.0.1

# Image downloading
aiohttp==3.9.3
aiofiles==23.2.1
